Convert IMU data processing to float

+++ pico_southbridge/driver/mpu6050/mpu6050.hpp
@@ -15,9 +15,9 @@ typedef struct _axles_int16 {
 } axles_int16;
 
 typedef struct _axles_float {
-  int16_t x;
-  int16_t y;
-  int16_t z;
+  float x;
+  float y;
+  float z;
 } axles_float;
 
 class mpu6050 {
